rename : rider get response 위치 변경
