+++ match-service/src/main/java/com/ohdelivery/service/match/rider/application/dto/response/GetRiderResponse.java
@@ -1,4 +1,4 @@
-package com.ohdelivery.service.match.rider.presentation;
+package com.ohdelivery.service.match.rider.application.dto.response;
 
 import java.util.UUID;
 import lombok.Getter;
